requirements_rag_enhancement.txt
# Dependências para aprimoramento do RAG
# Execute: pip install -r requirements_rag_enhancement.txt

# Core AI/ML
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# Text processing
tiktoken>=0.5.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web scraping
requests>=2.31.0
urllib3>=2.0.0

# Data handling
pandas>=2.0.0

# Caching (opcional)
redis>=5.0.0

# Async support
aiohttp>=3.8.0
aiofiles>=23.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0

# Development/Testing (opcional)
pytest>=7.4.0
pytest-asyncio>=0.21.0